+++ CHANGELOG.rst
@@ -11,6 +11,10 @@ Unreleased
 Added
 -----
 
+- State constraints on arbitrary fragments, targeting the electronic ground
+  state by determining a self-consistent constraint potential
+  (restricted to Mulliken populations at the moment)
+
 
 Changed
 -------